backend/requirements.txt
asgiref==3.6.0
Django==4.2
psycopg2-binary==2.9.6
python-dotenv==1.0.0
sqlparse==0.4.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
requests==2.28.2
django-filter==23.2